requirements.txt
Flask==2.3.3
Pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.26.0
gunicorn==21.2.0
